move version() to SVMStaticMessage trait
