+++ cost-model/src/transaction_cost.rs
@@ -191,6 +191,10 @@ impl WritableKeysTransaction {
 
 #[cfg(feature = "dev-context-only-utils")]
 impl solana_svm_transaction::svm_message::SVMStaticMessage for WritableKeysTransaction {
+    fn version(&self) -> solana_transaction::versioned::TransactionVersion {
+        unimplemented!("WritableKeysTransaction::version")
+    }
+
     fn num_transaction_signatures(&self) -> u64 {
         unimplemented!("WritableKeysTransaction::num_transaction_signatures")
     }
@@ -268,10 +272,6 @@ impl solana_svm_transaction::svm_message::SVMMessage for WritableKeysTransaction
 
 #[cfg(feature = "dev-context-only-utils")]
 impl solana_svm_transaction::svm_transaction::SVMTransaction for WritableKeysTransaction {
-    fn version(&self) -> solana_transaction::versioned::TransactionVersion {
-        unimplemented!("WritableKeysTransaction::version")
-    }
-
     fn signature(&self) -> &solana_signature::Signature {
         unimplemented!("WritableKeysTransaction::signature")
     }
