+++ runtime-transaction/src/runtime_transaction.rs
@@ -83,6 +83,10 @@ impl<T> Deref for RuntimeTransaction<T> {
     }
 }
 impl<T: SVMStaticMessage> SVMStaticMessage for RuntimeTransaction<T> {
+    fn version(&self) -> TransactionVersion {
+        self.transaction.version()
+    }
+
     fn num_transaction_signatures(&self) -> u64 {
         self.transaction.num_transaction_signatures()
     }
@@ -165,10 +169,6 @@ impl<T: SVMMessage> SVMMessage for RuntimeTransaction<T> {
 }
 
 impl<T: SVMTransaction> SVMTransaction for RuntimeTransaction<T> {
-    fn version(&self) -> TransactionVersion {
-        self.transaction.version()
-    }
-
     fn signature(&self) -> &Signature {
         self.transaction.signature()
     }
