+++ svm-transaction/src/svm_transaction/sanitized_transaction.rs
@@ -1,19 +1,9 @@
 use {
-    crate::svm_transaction::SVMTransaction,
-    solana_message::SanitizedMessage,
-    solana_signature::Signature,
-    solana_transaction::{sanitized::SanitizedTransaction, versioned::TransactionVersion},
+    crate::svm_transaction::SVMTransaction, solana_signature::Signature,
+    solana_transaction::sanitized::SanitizedTransaction,
 };
 
 impl SVMTransaction for SanitizedTransaction {
-    fn version(&self) -> TransactionVersion {
-        match self.message() {
-            SanitizedMessage::Legacy(_) => TransactionVersion::LEGACY,
-            SanitizedMessage::V0(_) => TransactionVersion::Number(0),
-            SanitizedMessage::V1(_) => TransactionVersion::Number(1),
-        }
-    }
-
     fn signature(&self) -> &Signature {
         SanitizedTransaction::signature(self)
     }
